separated db operations into db_utils submodule
